Added MainMenu, PauseMenu Object and EndMenu
Several asset files were added including a pauseMenu prefab (requires an event system to be in the room for functionaility). as well as some sprites for trees and rocks, a MenuSoundEffect and MenuParticles.
Also includes scenes for the MainMenu and WinScreen menu with functionality pretty much complete (some functions can only be used one we have more content in the game)